migrating to helper js file
